Extract TrelloAuthentication into trello_authentication.dart
