+++ lib/src/sdk/cards/attachment_client.dart
@@ -1,7 +1,7 @@
-import '../client.dart' show TrelloAuthentication;
 import '../file_name.dart' show FileName;
 import '../http_client.dart' show HttpClient;
 import '../query_options.dart' show QueryOptions;
+import '../trello_authentication.dart' show TrelloAuthentication;
 import 'cards.dart'
     show AttachmentFields, AttachmentId, CardId, TrelloAttachment;
