+++ lib/trello_sdk.dart
@@ -11,5 +11,6 @@ export 'src/sdk/http_client.dart';
 export 'src/sdk/lists/lists.dart';
 export 'src/sdk/members/members.dart';
 export 'src/sdk/misc.dart';
+export 'src/sdk/trello_authentication.dart';
 export 'src/sdk/trello_models.dart';
 export 'src/sdk/trello_object.dart';
